Add a "squash_time_gaps=y" option to MS transposition.
This gets rid of runs where there are no data; this can happen if, for
instance, you're looking at data from periodic calibrator visits that have
large intervening gaps from science observations. Because we can only support
linear time axes, this means that the timestamps for all but the first chunk
of data will be invalid. But that's not always a dealbreaker.