scanalyzer/transpose.py: fix LSB spectral windows
I had tried to make things work when the dataset contained a lower-sideband
spectral window, which is usually indicated by having its frequency channels
have a negative separation. But I messed it up: the code copied the data into
the buffers with proper reversal of the array ordering, but it still started
the copy at the index of the zeroth channel in the SPW, when it should
actually have started the copy at the lowest-frequency channel, i.e. the last
channel. Now fixed.